Add support for the old style handshake
The handshake process now automatically detects whether to use the new style
or the old style.  For the new style handshake, still only the fixed handshake
is supported.

This commit also includes a workaround for the Scaleway servers claiming to
support TRIM but then returning EINVAL when a trim command is attempted.